LayoutTests/inspector/view-css-expected.txt

This tests that CSS files are properly loaded / unloaded upon view show / detach.


Running: testEvents
Parent.show()
  LOAD parent.css
Child 1.show()
  LOAD child.css
  LOAD child1.css
Child 2.show()
  LOAD child2.css
Child 2.detach()
  UNLOAD child2.css
Parent.detach()
  UNLOAD parent.css
  UNLOAD child.css
  UNLOAD child1.css
Parent.show()
  LOAD parent.css
  LOAD child.css
  LOAD child1.css
